requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
build==1.2.2.post1
    # via pip-tools
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.54.1
    # via matplotlib
gecatsim==1.4.4
    # via -r requirements.in
kiwisolver==1.4.7
    # via matplotlib
matplotlib==3.9.2
    # via
    #   gecatsim
    #   vtk
numpy==2.1.2
    # via
    #   -r requirements.in
    #   contourpy
    #   gecatsim
    #   matplotlib
    #   scipy
    #   tifffile
    #   vedo
packaging==24.1
    # via
    #   build
    #   matplotlib
pillow==11.0.0
    # via matplotlib
pip-tools==7.4.1
    # via -r requirements.in
pydicom==3.0.1
    # via -r requirements.in
pygments==2.18.0
    # via vedo
pyparsing==3.2.0
    # via matplotlib
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via matplotlib
scipy==1.14.1
    # via gecatsim
six==1.16.0
    # via python-dateutil
tifffile==2024.9.20
    # via tiffile
tiffile==2018.10.18
    # via -r requirements.in
tqdm==4.66.6
    # via gecatsim
typing-extensions==4.12.2
    # via vedo
vedo==2024.5.2
    # via -r requirements.in
vtk==9.3.1
    # via vedo
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
